Register "Extra->Play-at-Speed"; break TransportMenus->...
...TranscriptionToolBar
